ai-matching-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg[binary]==3.2.9
psycopg-pool==3.2.4
pydantic==2.11.7
python-dotenv==1.1.1
sentence-transformers==3.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-mock==3.14.0
